requirements.txt
# Hydra-Map — Pinned Minimal Dependencies (Windows-compatible)
# Install: pip install -r requirements.txt
#
# NOTE: 'segment-anything' requires git to be on PATH.
#   Alternative manual install (if git fails):
#     pip install git+https://github.com/facebookresearch/segment-anything.git
#   Or skip it — GeoSAM will run in stub mode automatically.

# Core ML
torch>=2.0.0
torchvision>=0.15.0
timm>=0.9.0
albumentations>=1.3.0

# YOLO
ultralytics>=8.0.0

# Segment Anything (GeoSAM) — requires git on PATH
# If this fails, comment it out; the pipeline will use stub mode automatically.
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git

# Geospatial
rasterio>=1.3.0
geopandas>=0.13.0
shapely>=2.0.0
fiona>=1.9.0
pyproj>=3.5.0

# Point cloud
laspy[laszip]>=2.5.0

# ONNX / Edge
onnx>=1.14.0
onnxruntime>=1.15.0

# Utilities
numpy>=1.24.0
scipy>=1.10.0
scikit-learn>=1.2.0
opencv-python-headless>=4.7.0
Pillow>=9.5.0
pyyaml>=6.0
tqdm>=4.65.0

# Observability (optional — disabled by default in config.yaml)
# wandb>=0.15.0

# Testing
pytest>=7.3.0
